+++ crates/ui-components/src/lib.rs
@@ -40,6 +40,7 @@ pub mod code;
 pub mod code_block;
 pub mod color_area;
 pub mod color_field;
+pub mod color_picker;
 pub mod color_slider;
 pub mod color_swatch;
 pub mod color_swatch_picker;
@@ -189,6 +190,7 @@ pub use code::{Code, CodeVariant};
 pub use code_block::{CodeBlock, CodeBlockMotion};
 pub use color_area::ColorArea;
 pub use color_field::ColorField;
+pub use color_picker::ColorPicker;
 pub use color_slider::{ColorSlider, ColorSliderChannel, ColorSliderMotion};
 pub use color_swatch::{
     ColorSwatch, ColorSwatchAlpha, ColorSwatchRounding, ColorSwatchShape, ColorSwatchSize,